pyproject.toml
[project]
name = "tesla-coil-controller"
version = "0.1.0"
description = ""
authors = [
    {name = "HetorusNL",email = "tim@hetorus.nl"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "mido (>=1.3.3,<2.0.0)",
    "pyserial (>=3.5,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
